refactor: extract flowrs-config as workspace crate
