+++ Cargo.toml
@@ -44,6 +44,7 @@ formula = "flowrs"
 [dependencies]
 ansi-to-tui = { version = "8.0.1" }
 anyhow = { workspace = true }
+flowrs-config = { path = "crates/flowrs-config" }
 async-trait = { workspace = true }
 aws-config = { version = "1.8.11" }
 aws-sdk-mwaa = "1.99.0"
