+++ crates/flowrs-config/Cargo.toml
@@ -8,3 +8,13 @@ license = "MIT"
 repository = "https://github.com/jvanbuel/flowrs"
 
 [dependencies]
+anyhow.workspace = true
+clap.workspace = true
+dirs.workspace = true
+log.workspace = true
+serde.workspace = true
+strum.workspace = true
+toml.workspace = true
+
+[lints.clippy]
+pedantic = "warn"
