+++ src/airflow.rs
@@ -1,5 +1,4 @@
 pub mod client;
-pub mod config;
 pub mod graph;
 pub mod managed_services;
 pub mod model;
