+++ src/airflow/client/v1/mod.rs
@@ -12,7 +12,8 @@ use reqwest::Method;
 use url::{form_urlencoded, Url};
 
 use super::base::BaseClient;
-use crate::airflow::{config::AirflowVersion, traits::AirflowClient};
+use flowrs_config::AirflowVersion;
+use crate::airflow::traits::AirflowClient;
 use crate::airflow::model::common::OpenItem;
 
 /// API v1 client implementation (for Airflow v2, uses /api/v1 endpoint)
