+++ src/airflow/managed_services/composer/auth.rs
@@ -1,1 +1,1 @@
-pub use crate::airflow::config::managed_auth::ComposerAuth;
+pub use flowrs_config::managed_auth::ComposerAuth;
