+++ src/airflow/managed_services/mwaa.rs
@@ -1,5 +1,6 @@
 use crate::airflow::client::auth::AuthProvider;
-use crate::airflow::config::{AirflowAuth, AirflowConfig, AirflowVersion, ManagedService};
+use flowrs_config::{AirflowAuth, AirflowConfig, AirflowVersion, ManagedService};
+use flowrs_config::managed_auth::{MwaaAuth, MwaaTokenType};
 use anyhow::{Context, Result};
 use async_trait::async_trait;
 use aws_config::BehaviorVersion;
@@ -7,8 +8,6 @@ use aws_sdk_mwaa as mwaa;
 use log::info;
 use reqwest::RequestBuilder;
 use serde::Serialize;
-
-use crate::airflow::config::managed_auth::{MwaaAuth, MwaaTokenType};
 
 /// MWAA client for managing authentication and environment discovery
 #[derive(Debug, Clone)]
