+++ src/app/model/filter/impls.rs
@@ -3,7 +3,7 @@
 //! This module centralizes all `Filterable` implementations, keeping them
 //! separate from the domain model definitions.
 
-use crate::airflow::config::AirflowConfig;
+use flowrs_config::AirflowConfig;
 use crate::airflow::model::common::dag::Dag;
 use crate::airflow::model::common::dagrun::DagRun;
 use crate::airflow::model::common::taskinstance::TaskInstance;
