+++ src/app/state.rs
@@ -1,6 +1,6 @@
 use std::sync::LazyLock;
 
-use crate::airflow::config::FlowrsConfig;
+use flowrs_config::FlowrsConfig;
 use crate::airflow::model::common::{DagId, DagRunId, TaskId};
 use crate::app::environment_state::EnvironmentStateContainer;
 use crate::app::model::dagruns::DagRunModel;
