+++ src/commands/config/update.rs
@@ -5,10 +5,8 @@ use log::info;
 use strum::IntoEnumIterator;
 
 use super::model::UpdateCommand;
-use crate::{
-    airflow::config::{AirflowAuth, AirflowConfig, BasicAuth, FlowrsConfig, TokenSource},
-    commands::config::model::{validate_endpoint, ConfigOption},
-};
+use flowrs_config::{AirflowAuth, AirflowConfig, BasicAuth, FlowrsConfig, TokenSource};
+use crate::commands::config::model::{validate_endpoint, ConfigOption};
 
 use anyhow::Result;
